drizzle-next-neon-clerk

+++ drizzle/0000_serious_whirlwind.sql
@@ -1,5 +1,5 @@
 CREATE TABLE "todos" (
-	"id" bigint PRIMARY KEY NOT NULL,
+	"id" serial PRIMARY KEY NOT NULL,
 	"text" text NOT NULL,
 	"done" boolean DEFAULT false NOT NULL,
 	"user_id" integer DEFAULT 1
